refactor: 数据库连接优化
